# backend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediSync API",
    "description": "Hospital Management System API for managing doctors, patients, and appointments.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/allergies/{allergyId}": {
      "get": {
        "tags": [
          "Allergies"
        ],
        "summary": "Get allergy by ID",
        "description": "Retrieves details for a specific allergy.",
        "operationId": "getAllergy",
        "parameters": [
          {
            "name": "allergyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved allergy",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AllergyDto"
                }
              }
            }
          },
          "404": {
            "description": "Allergy not found"
          }
        }
      },
      "put": {
        "tags": [
          "Allergies"
        ],
        "summary": "Update allergy",
        "description": "Updates details of an existing allergy. Requires ADMIN role.",
        "operationId": "updateAllergy",
        "parameters": [
          {
            "name": "allergyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllergyCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Allergy updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AllergyDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role"
          },
          "404": {
            "description": "Allergy not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Allergies"
        ],
        "summary": "Delete allergy",
        "description": "Removes an allergy from the system. Requires ADMIN role.",
        "operationId": "deleteAllergy",
        "parameters": [
          {
            "name": "allergyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Allergy deleted successfully"
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role"
          },
          "404": {
            "description": "Allergy not found"
          }
        }
      }
    },
...
    "/api/allergies": {
      "get": {
        "tags": [
          "Allergies"
        ],
        "summary": "List all allergies",
        "description": "Retrieves the list of all defined allergies in the system.",
        "operationId": "getAllAllergies",
        "responses": {
          "200": {
            "description": "Successfully retrieved list",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AllergyDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Allergies"
        ],
        "summary": "Create new allergy",
        "description": "Adds a new allergy to the system. Requires ADMIN role.",
        "operationId": "createAllergy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllergyCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Allergy created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AllergyDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (e.g. missing name or duplicate code)"
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role"
          }
        }
      }
    },
    "/api/admins": {
      "get": {
        "tags": [
          "Admin Management"
        ],
        "summary": "List all admins",
        "description": "Retrieves a list of all users with ADMIN role.",
        "operationId": "getAllAdmins",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin Management"
        ],
        "summary": "Register new admin",
        "description": "Creates a new system administrator account.",
        "operationId": "registerAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRegistrationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Admin created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or email already exists"
          }
        }
      }
    },
...
    "/api/admins/{adminId}": {
      "delete": {
        "tags": [
          "Admin Management"
        ],
        "summary": "Delete admin",
        "description": "Removes an administrator account. NOTE: You cannot delete yourself.",
        "operationId": "deleteAdmin",
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Admin deleted successfully"
          },
          "400": {
            "description": "Bad Request (e.g., trying to delete yourself)"
          },
          "404": {
            "description": "Admin not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AllergyCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "code": {
            "type": "string",
            "minLength": 1
          },
          "category": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "category",
          "code",
          "name"
        ]
      },
      "AllergyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "category": {
            "type": "string"
          }
        }
      },
...
      "AdminRegistrationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "pattern": "^(?=.*[0-9])(?=.*[A-Z])(?=.*[!@#$%^&()+=_-])(?=\\S+$).{8,}$"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}